A0-2541: Manually check dependat job statuses in check e2e job completion (#1243)
# Description

Surprisingly GitHub treats skipped jobs status as success, despite those
skipped jobs are Required Checks. With the introduction of Merge Queues,
this has a danger to sneak in a change that fails one of e2e tests. See
all compiled info in this
[answer](https://github.com/orgs/community/discussions/46757#discussioncomment-5831752)

## Type of change

Please delete options that are not relevant.

- Bug fix (non-breaking change which fixes an issue)

# Checklist:

* Tested this change in testing repo
https://github.com/gen64/repo-for-testing/pull/28/files