A0-2088: Better task system in sync (#1212)
# Description

The task system in sync was sufficient for justification requests, but
now that we are moving on to actually requesting blocks we need
something slightly better. Otherwise we would be susceptible to an
attack that would stop us from requesting some blocks given "fake"
headers provided in a specific order.

Also improves the forest interfaces to make things less `mut`.

## Type of change

- Bug fix (non-breaking change which fixes an issue)

# Checklist:

- I have added tests
- I have made corresponding changes to the existing documentation
- I have created new documentation

---------

Co-authored-by: timorl <timorl@disroot.org>