A0-2325: Rename RPC endpoint (#1250)
# Description

We rename RPC endpoint from `chain_getBlockAuthor` to
`alephNode_getBlockAuthor`. This is because polkadotJS library doesn't
support extending `chain` (in general: any default) RPC method section.

I have also made use of `namespace` attribute of `rpc` macro to clean
the code a bit.

In future we can easily add an alias `chain_getBlockAuthor`, but for now
it seems to be unnecessary (consulted with Damian).

## Type of change

- Bug fix (non-breaking change which fixes an issue)