A0-3083: Refactor ticker for chain extensions (#1367)
# Description

These are changes that are needed to use ticker with the periodic
requests for chain extensions which we will be implementing soon. In
particular this _should not_ change any current behaviour of the node,
thus it's just a refactor.

Comments on whether this interface is clear or not are very welcome.

## Type of change

- New feature (non-breaking change which adds functionality)

# Checklist:

- I have added tests
- I have made corresponding changes to the existing documentation
- I have created new documentation

---------

Co-authored-by: timorl <tomasz.kisielewski@cardinals.cc>