A0-3266: Remove updatenet workflows (#1425)
# Description

We have decided to remove Updatenet workflows as they bring a lot of
confusion. The name first appeared in one of the ADR's. Updatenet is
basically a featurenet that is started from an existing net (testnet or
mainnet), not branch.
Hence, any temporary aleph-node networks from now on are just
Featurenets. There would be too much work to rename that to "temporary"
or anything more general.

Current workflows will just have `FE` in their display name as it's
common to everyone that it means a featurenet. Why waste space, right?

Also, we have moved featurenet workflows to the external
`github-actions` repository where we keep all shared github actions
code. Hence, the current workflows will calls these external guys. That
was always something that had to be done, and now we've got a case where
other projects/repository need to have their own featurenets, with
separate workflows. So it's basically to keep things DRY.

What is worth mentioned as well is that, we have automatic tests for
featurenets now.