A0-3430: Update ink deps to latest ink4-compatible releases (#1463)
# Description

Updates ink-releated dependencies in `aleph-client`. There were no
changes in the API or in the code, purely dependency bump. This update
is necessary to align `aleph-client` with `ink-wrapper` which works with
`ink` and `aleph-client` as a backend.

The versions of dependencies used are going to be the last compatible
with ink4


## Type of change

- Dependency versions update.

# Checklist:

- I have bumped aleph-client version
- I have bumped cliain verison