A0-3454: All e2e tests should be run also on release branches (#1467)
# Description

In change
https://github.com/Cardinal-Cryptography/aleph-node/commit/ad530be7c0035c803e100d2289ba224a01566f73
all e2e tests except finalization have been moved to nightly. But we
need to automatically run them on the release branch as there are no
nightly release branches. There are seldom commits on release branches
so there should not be a problem for costs/developer synchronization.

Also, check runtime determinism is important to run, especially on the
release branches. Since is more important to run it on every release
candidate, not just nightly, I moved it to be run for every push to
`main` or release branch. Hence, we do not need nightly check runtime
determinism anymore.

Long integration tests like `clique network that are on nightly should
be run on release branch as well.

## Type of change

Please delete options that are not relevant.

- New feature (non-breaking change which adds functionality)