A0-3432: run_nodes.sh revamp, in particular run bootnode as RPC node (#1472)
# Description

This PR revamps `scripts/run_nodes.sh` so that
* there are only two categories of nodes: RPC node and validator node
* boot node is always the first RPC node
* number of either node type is configurable, but sum must not exceed
10,
* script cmd line args was reworked which is a breaking change
* by default, script ideas all the following: builds `aleph-node`,
bootstraps chain, purges db, removes AlephBFT backups. User can opt-out
from either of these, but obviously some of them must be present if
others are etc, so the script checks all that
* script now logs with timestamps and colors, which is justifies as the
script is not intended to run in CI
* as the script is a front card of our repo and most likely the first
point of contact with the new user, it contains now a compact intro
to chain bootstrap in general
* `jq` is now needed, though is not hard blocker for me to revert to
previous behaviour, now it is just cleaner in bash to parse output of
`aleph-node key inspect`
* base path default was changed from `/tmp/` to local `./run-nodes-sh`,
this is because the user must have write permissions in the repo he or
she just cloned, and it's better to have a base path that is not shared
with any other files/dirs.
* arguments were synced with `docker/docker_entrypoint.sh`; they are
indeed duplicated now but it does not make sense to extract them to
common .sh as at some point `run_nodes.sh` might become obsolete due to
dockerization

## Type of change

Please delete options that are not relevant.

- New feature (non-breaking change which adds functionality)
- Breaking change (fix or feature that would cause existing
functionality to not work as expected)