A0-3143: Major sync test does not use pre-baked workdir. (#1481)
# Description

This PR removes pre-baked `workdir` file used in major sync test.
Motivations are:
* that `workdir` was generated on 11.X version. It has happened on the
`main` branch in which there was a change that changed database schema
in a way that finalization does not work. This test is supposed to
validate whether node is able to sync in major sync mode, ie when gap is
greater then 2 sessions, and not major update scenario which it does now
* `workdir` contributes to 20% of whole `aleph-node` size and it bloats
git history
* as it is now, we need to regenerate `workdir` every time such breaking
change occurrs, after this PR is merge we won't have to do it

Also, this PR removes some tech debt which is:
* removed not used code for python tests, which functionality is
covererd by major sync and force reorgs tests
* added common bash script for running python tests

## Type of change

Please delete options that are not relevant.

- Bug fix (non-breaking change which fixes an issue)
- New feature (non-breaking change which adds functionality)

## Testing

- normal session nightly pipeline pass
https://github.com/Cardinal-Cryptography/aleph-node/actions/runs/6798163085/job/18482394829