A0-3432: E2E tests run now with RPC node as bootnode. (#1476)
# Description

Following up
https://github.com/Cardinal-Cryptography/aleph-node/commit/21ead16619c24e8e5197e8de1c368905a604141a,
we introduce RPC Node to e2e docker-compose framework. This is main, non
breaking change to anybody apart from e2e package crate system (CI).
Other than that, below fixes have been made:
* `run_consensus.sh` now runs only amount of nodes user passes, now all
nodes (8) in the common docker compose file, to not waste CPU/memory
* RPC_PORT now it's 9944 - this was last place it was non-unified,
* setting minimum validator count is not required, meaning that from e2e
test perspective it does not matter if we pass 7 as currently we do or
it is default 4 value - tests passes anyway. Hence, we removed that
feature
* `run_consensus.sh` now can be run from any path,
* common docker-compose now supports up to 9 validators
* stash now equals to controller in chain_spec generation. This code was
dead anyway as substrate 1.0.0 skips controller anyway:
https://github.com/Cardinal-Cryptography/substrate/blob/aleph-v1.0.0/frame/staking/src/pallet/mod.rs#L618
* removed not used `RANDOMIZED` feature in e2e tests
* updated sync and latency tests to point to 9944 port

## Type of change

Please delete options that are not relevant.

- Bug fix (non-breaking change which fixes an issue)
- New feature (non-breaking change which adds functionality)
- Breaking change (fix or feature that would cause existing
functionality to not work as expected)

## Testing:

- nightly e2e logic tests pass:
https://github.com/Cardinal-Cryptography/aleph-node/actions/runs/6783539788
- normal session e2e tests: pass
https://github.com/Cardinal-Cryptography/aleph-node/actions/runs/6795064373