A0-2800: Migrate button contracts (#1529)
# Description

This PR migrates all contracts from `aleph-node` except `adder` to a
separate repo. Those are all essentially single product that is called
`the-button` app. All CI, including button e2e test was removed - we
need to fix them first, if ever touched, only then add CI to the
`the-button` repo.

Note that button tests in `aleph-e2e-client` were not removed, this was
on purpose - this code has the potential to be called in the future and
a lot of time was invested in it already, in contrast to the button CI
part which should be written from the beginning in the new repo, if
required.

Apart from that, the following tech debt was removed:
* instead of invoking `cargo contracts install`, `ink-dev` docker is
used. This saved 4-5 minutes in `_check-excluded-packages` workflow, and
around 2-3 minutes in total `PR commit` worfklow.

Related PRs:
* https://github.com/Cardinal-Cryptography/the-button/pull/134
* https://github.com/Cardinal-Cryptography/aleph-node/pull/1526
* https://github.com/Cardinal-Cryptography/aleph-node/pull/1307 - this
will be closed as-is

## Type of change

Please delete options that are not relevant.

- New feature (non-breaking change which adds functionality)