A0-3618: move proof verification from pallet to chain extension (#1536)
# Description

As decided recently, we don't need verification in the pallet. It is
sufficient to have directly in the chain extension.

Since we are removing jellyfish soon, I didn't migrate the verifier
tests.

Technically this is a breaking change for aleph-client and cliain, but
actually it's not used anywhere, hidden behind a feature flag and there
will be more changes like this soon, and thus I bumped just the minor
version.

## Type of change
- Breaking change (fix or feature that would cause existing
functionality to not work as expected)

# Checklist:

- I have made neccessary updates to the Infrastructure
- I have made corresponding changes to the existing documentation
- I have created new documentation
- I have bumped aleph-client version if relevant