A0-3618: `VkStorage` pallet (#1541)
# Description

The last part of the deep cleaning of the ZK code. We rename
`baby-liminal-pallet` to `vk-storage-pallet` as now it only serves as a
dummy, persistent storage for verification keys.

Instead of using `<T as frame_system::Config>::Hashing` I have just
hardcoded Blake256 hashing, so that the chain extension is not generic
over runtime configuration.

Pallet benchmarks have been rerun for the new (and only) extrinsic.

## Type of change

- Breaking change (fix or feature that would cause existing
functionality to not work as expected)

# Checklist:

<!-- delete when not applicable to your PR -->

- I have added tests
- I have made neccessary updates to the Infrastructure
- I have made corresponding changes to the existing documentation
- I have created new documentation