A0-3835: Halo2 verifier (#1561)
# Description

We add halo2 SNARK verification as a runtime interface. This is the last
piece in the whole pipeline that enables proof verification in smart
contracts. Now, the flow is:

```
 Smart Contract -> Chain Extension -> Runtime interface API -> Runtime interface implementation
|--- SC side ---|---------------- runtime ------------------|------------- host----------------|
```

I put runtime interface into a new crate, so that we keep `primitives`
as simple as possible. It can serve for any future runtime interfaces as
well. It was put outside workspace, similarly to `pallet-vk-storage` and
`baby-liminal-chain-extension`, for the similar reason.

Notes:
- to adapt to the halo2 verification API and avoid changing core code,
I'm using mocked, lightweight KZG parameters
- there is a public constant `CIRCUIT_MAX_K` which describes (lg of) the
maximum number of rows in the supported circuits; for now it is
ridiculously low (12, meaning that we can have up to ~4096 rows), so
that testing is fast (every time we want to generate proving/verifying
keys we actually have to compute `2^CIRCUIT_MAX_K` group elements); I
will spend some time on it in the next week (static cache for tests,
maybe some simplifications)
- `DeserializingProofFailed`-type error no longer applies to
verification process; proof deserialization happens inside
`verify_proof` method and will be translated to a more general variant
of `VerificationFailed`
- companion PR:
https://github.com/Cardinal-Cryptography/pse-halo2/pull/1

## Type of change

- New feature (non-breaking change which adds functionality)

# Checklist:

<!-- delete when not applicable to your PR -->

- I have added tests
- I have made neccessary updates to the Infrastructure
- I have made corresponding changes to the existing documentation
- I have created new documentation