A0-3849: Implicitly expect polynomial degree bound within verification key encoding (#1566)
# Description

Follow-up to
https://github.com/Cardinal-Cryptography/aleph-node/pull/1561#discussion_r1442786888

We expect that the keys stored in the pallet contain `k` - logarithm of
the polynomial degree bound (a parameter of the trusted setup). Thanks
to this we can get rid of the hard-coded, common constant in the
verifier code.

Note: in the current halo2 implementation, verifying key indirectly
contains `k` (through its evaluation domain field), but this is a
low-level, internal detail, that we don't want to depend on.

## Type of change

Please delete options that are not relevant.

- New feature (non-breaking change which adds functionality)

# Checklist:

- I have made neccessary updates to the Infrastructure
- I have made corresponding changes to the existing documentation
- I have created new documentation