A0-3759: Replace `HeaderBackend`, `BlockchainEvents` and `SelectChain` (#1554)
# Description

To remove completely substrate's `BlockT` abstraction, we have to
replace all substrate's abstraction parametrized with it that we use.
These mostly includes `ClientForAleph` supertraits. This PR creates our
abstraction for the traits mentioned in the title.

This PR moves conflicting areas, that have to use both abstractions, to
higher level code (namely `party`, see `NodeSessionManagerImpl`). We
still have to somehow wrap `AlephSessionApi`, so that its `BlockT` usage
is hidden, but I think this is out of the scope of this PR. The same
goes with aggregator, that currently uses `Justification` from
`block::substrate`, instead of the generic one.

Also, the whole `testing` directory is kind of problematic, containing
tests that are neither unit not purely integration ones. We might want
to revamp them a bit in the future to make them of either of these,
instead of something in between. Previously, we've used `TBlock`, that
was differing with the production block type only by block extrinsic
type. I changed these tests to use pure production block type (differs
in extrinsic). In the future, we might replace it with `MockBlock`.

## Type of change

None - refactor

# Checklist:

<!-- delete when not applicable to your PR -->

- I have made corresponding changes to the existing documentation
- I have created new documentation