A0-3861:  Increase MaxCodeLen for contracts. (#1570)
# Description

THIS IS A RUNTIME-ONLY CHANGE.

Increases the maximum code len for a contract from 128 KB to 256 KB. The
reason this is safe is that the constant 72 that we have in our
integrity test changed:
- first 72 -> 68 (which we didn't mirror in our repo, but it's here
https://github.com/paritytech/polkadot-sdk/blob/06fa111f58b0172c9f4ba7c209c8b6945787faeb/substrate/frame/contracts/src/lib.rs#L439
) this was a result of changing gas metering
- then 68 -> 32 -- this is a result of updating wasmi to v0.30 which
reduces the amount of memory each instruction requires. In slack I
include a message from a Pallet Contract Developer which clarifies this
further.


# Release note:

- Increased maximum length of a contract from 128KB to 256 KB.