A0-3899: Our base protocol outline (#1581)
# Description

Adds an outline of the new base protocol, replacement for
`SyncingEngine`. Does not have the protocol, but does some of the other
things we will need.

I think some of the imports will have to be changed during the next
substrate updates, because the definitions are now in different crates,
but oh well.

## Type of change

- New feature (non-breaking change which adds functionality)

# Checklist:

- I have added tests
- I have created new documentation