A0-3858: Benchmark on-chain verifier (1/2) (#1583)
# Description

We run benchmarks for the on-chain verifier. 

Even though the actual verification happens in the
`aleph-runtime-interfaces` crate, we have to put them in the chain
extension library for two reasons:
1. we already have benchmarking setup ready for the CE and it is CE,
that is the only consumer of these metrics
1. additionally, we need to measure verification key access

For now, we use only a single circuit. The follow-up PR will extend
measurements for other layouts.

There was also a very nasty bug introduced in #1405 which turned off
custom host functions from the benchmarking node. It has been fixed in
this PR.

For the people who can read halo2 diagrams, a diagram of a circuit
<5,3>:

![image](https://github.com/Cardinal-Cryptography/aleph-node/assets/27450471/e584851e-2311-4667-84c0-64f893beb564)


## Type of change

- Bug fix (non-breaking change which fixes an issue)
- New feature (non-breaking change which adds functionality)

# Checklist:

- I have made neccessary updates to the Infrastructure
- I have made corresponding changes to the existing documentation
- I have created new documentation