A0-3858: Benchmark on-chain verifier (2/2) (#1600)
# Description

Follow-up to #1583:

1. Make `BenchCircuit` non-generic: it posed ridiculous problems when
trying to run proof generation for a matrix of parameters.
2. Run trusted setup for `k = 20` (it took ~2h on `12th Gen Intel® Core™
i7-12800H × 20` with `16GB` RAM)
3. Run actual benchmarks for a cartesian product of parameters:
  - number of instances from `[1, 2, 8, 16, 64, 128]`
- gate blowup from `[1, 8, 64, 512, 4096]` (maximum number of rows was
`128 · 4096 ~= 0.5M`

To avoid awful boilerplate with 30 repeated functions in the benchmark
suite, I wrote a semi-awful macro.

I added all the artifacts to VCS, but unfortunately I forgot to save
SRS... :see_no_evil: in case we need to rerun benches, we have to
remember about it :stuck_out_tongue:

## Type of change

Please delete options that are not relevant.

- New feature (non-breaking change which adds functionality)

# Checklist:

- I have made neccessary updates to the Infrastructure
- I have made corresponding changes to the existing documentation
- I have created new documentation