A0-4172: Fix for FE e2e nightly (#1653)
# Description

When an ECR image exists, this workflow is a no-op and silently skips
e2e test run. This is because GHA skips a job when any parent job is
skipped.

Unfortunately, there is no way to test this as the scenario in which bug
occurs is when there's already an image on ECR, ie when e.g. running
this workflow from `main`. So I've just run workflow from this PR
https://github.com/Cardinal-Cryptography/aleph-node/actions/runs/8292896136,
and if its okay then we can merge this PR and re-run from main to see if
it helped.

## Type of change

Please delete options that are not relevant.

- Bug fix (non-breaking change which fixes an issue)