A0-4216: Does not depend on actual runtime API in aleph-node: const and types refactor (#1666)
# Description

This PR moves some types from `aleph-runtime` to `primitives`. Also, it
cleans up and documents most of the types and consts in `primitives`. I
used
https://github.com/Cardinal-Cryptography/polkadot-sdk/blob/aleph-v1.4.0/polkadot/node/service/src/lib.rs
as a baseline for types, their order in the file and comments.

What can't be extracted though is `Block` type and its dependees. This
is because, in `primitives` it uses `OpaqueExtrinsic`, and in runtime
non-opaque one, so there are different types.

This is PR 1/2 of incorporating
[`fake_runtime_api.rs`](https://github.com/Cardinal-Cryptography/polkadot-sdk/blob/aleph-v1.4.0/polkadot/node/service/src/fake_runtime_api.rs)
into `aleph-node`.

---------

Co-authored-by: Grzegorz Gawryał <31205678+ggawryal@users.noreply.github.com>