A0-4231: Use json chainspec builder. Move chainspec generation to separate module. (#1683)
# Description

This is 1 out of N (2 or 3) PRs which goal is to remove `aleph-runtime`
dependency from `aleph-node`.

This PR:
* uses `json` chainspec builder (introduced in `polkadot 1.4.0`)
* moves chainspec generation to a separate module. Idea here is that
this module eventually becomes new crate.

Latter has bigger design goal, which is to move below commands to
separate bin crate called `generate-chain-spec`, ie moving
`bootstrap-chain` command out of `aleph-node` binary. This will be a
breaking change, but since it will be out in 14 release, with proper
communication should be fine, as it mostly affects new chains - ie dev
and testing ones.

The architecture of `Command` and new module files might not be ideal,
but will be improved in the subsequent PRs.

# Testing

- run e2e nightly
- compare manually generated non-raw chainspec with `main` version