A0-4231: Refactor all keystore and chainspec generation to seprate module (#1687)
# Description

Folow up after
https://github.com/Cardinal-Cryptography/aleph-node/commit/5a8b1c380fe3ebae863ba8c89ef4aa73b01d0d16

This is 2nd out of 3 PRs to remove `aleph-runtime` dependency in
`aleph-node`. In this PR:
* The whole keystore and chainspec generation logic is moved to
`chain_spec` module. This module will be extracted as a whole, with
small adjustment, to a separate crate
* `bootstrap-chain` command no longer supports `--faucet-account`
switch. This is because only reason for it to exists was to set an
initial endowment, exactly what `--rich-accounts` does. So it's
suggested to migrate all `--faucet-accounts <account-id>` to
`--rich-accounts <account-id>`. This is a breaking change, yet no
impactful it's not used frequently.
* `bootstrap-node` command is removed. This is because it was doing the
same what `bootstrap-chain` does, which now have
`--authorities-account-ids` switch to differentiate between RPC and
validator nodes

## Type of change

Please delete options that are not relevant.

- New feature (non-breaking change which adds functionality)
- Breaking change (fix or feature that would cause existing
functionality to not work as expected)

# Checklist:

* e2e tests:
https://github.com/Cardinal-Cryptography/aleph-node/actions/runs/8706331881