A0-3461: Update ABFT and remove legacy data IO (#1692)
# Description

Updates AlephBFT, letting us get rid of the ugly `data_io::legacy` hack.
Note that this updates `aleph_bft_crypto` which, in principle could be a
compatibility issue. Fortunately, the change from 0.8 to 0.9 is only in
some minor interfaces, the internal representation and, more
importantly, encoding of all the types remains unchanged.

## Type of change

- ABFT update

# Checklist: