A0-4251: Move `sign_with_params` to a new trait in aleph client (#1700)
# Description

In #1602, we've added `sign_with_params` to `SignedConnectionApi` in
aleph-client to be able to specify nonce in lower level api for signing
transactions, without submitting. Normally, this could be done in
synchronous manner.
However, in Most bridge we hide signing keys in a separate enclave, and
implement `SignedConnectionApi` by a client talking to that enclave, to
be a proxy between that enclave and a user. This:
1. Complicates creating synchronous signing implementation for that
proxy, as a channel to the enclave is async,
2. Forces implementing that lower level method, when it is in fact not
needed in Most use case.

Therefore, I'm splitting that trait.
Also, in that PR I forgot to make `SubmittableExtrinsic` public, so I'm
doing it now.

## Type of change

Please delete options that are not relevant.

- Bug fix (non-breaking change which fixes an issue)

# Checklist:

<!-- delete when not applicable to your PR -->

- I have made corresponding changes to the existing documentation
- I have bumped `aleph-client` version