A0-4231: Extract chain bootstrap to a separate crate (#1699)
# Description

This final PR achieves the goal of removing the `aleph-runtime`
dependency from `aleph-node`. A new crate `chain-bootstraper` is made
from the current chain bootstrap logic in `aleph-node`.

The dependencies in `aleph-node` should be clearer now:
* `aleph-node` does not compile native runtime anymore via default
features build.
* `chain-bootstrapper` does compile both native and wasm runtime during
build.
The amount of packages that node compiles is about 150 less than it was
before. Depending on the machine, It's up to 40% compilation speedup for
a`aleph-node` binary. Here's local build, clean repo, on my laptop:

After change - `aleph-node` and `chain-bootstrapper` builds in the
similar time.
```
time ./scripts/run_nodes.sh -v 6 
2024-04-22 13:37:54:233 Starting ./scripts/run_nodes.sh
2024-04-22 13:37:54:234 Creating base path ./run-nodes-local if it does not exist
2024-04-22 13:37:54:236 Stopping all current node processes
2024-04-22 13:37:54:241 No aleph-node processes found.
2024-04-22 13:37:54:242 Building testing aleph-node binary (short session) and chainspec-generator binary.
   Compiling proc-macro2 v1.0.81
[...]
    Finished release [optimized] target(s) in 9m 02s
[...]
    Finished release [optimized] target(s) in 9m 31s
[...]

________________________________________________________
Executed in   18.59 mins    fish           external
   usr time  119.30 mins  874.00 micros  119.30 mins
   sys time    5.67 mins    0.00 micros    5.67 mins

```

Before change - note though the overall time is faster than before:
```
marcin@marcin-Latitude-3520 ~/g/aleph-node-3 (main)> time ./scripts/run_nodes.sh -v 6                                            (base) 
[...]

    Finished release [optimized] target(s) in 16m 44s

[...]
Executed in   16.77 mins    fish           external
   usr time  107.24 mins    0.00 micros  107.24 mins
   sys time    5.32 mins  791.00 micros    5.32 mins

```
This is because of `polkadot-sdk` crates architecture - we still need to
compile most of the crates in both node and bootstrapper. There's small
optimization in this PR to not build three polkadot-sdk crates with
default features, ie to disable `rocks-db` build when
`chain-bootstrapper` is built.

In Ci though, compilation should be faster as we can parallelize
`aleph-node` and `chain-bootstrapper` build, which this PR achieves so.
Most of the changes are workflows anyway.

## Type of change

Please delete options that are not relevant.

- Breaking change (fix or feature that would cause existing
functionality to not work as expected)

# Checklist:

* Deploy to devnet: - will be tested after merge to `main`.
* Nightly feature-gated builds:
https://github.com/Cardinal-Cryptography/aleph-node/actions/runs/8828449304
* Nightly pipeline e2e tests on featurenet: - FEs engine require small
change, and will be tested after merge to `main`.
* Nightly pipeline integration tests:
https://github.com/Cardinal-Cryptography/aleph-node/actions/runs/8828450804
* Nightly pipeline logic e2e tests:
https://github.com/Cardinal-Cryptography/aleph-node/actions/runs/8828452203
* Nightly pipeline normal session e2e tests:
https://github.com/Cardinal-Cryptography/aleph-node/actions/runs/8828453925
* On main push workflow (edited so it's no-op):
https://github.com/Cardinal-Cryptography/aleph-node/actions/runs/8828493822