A0-2980: Fix sync handler rejecting proper blocks. (#1719)
# Description

When a block comes as a response to a request, we first process its
header, similarly as we would get this header in a response.

Additionally, next to making the `forest.importable(header.id)` check in
`handle_header` which checks if forest would accept the corresponding
block, we also allow for imports in a sequence. This means that if in
the same response processing we imported the parent of the block, then
we can also import the block, no matter what `forest.importable(...)`
has returned.

## Type of change

- Bug fix (non-breaking change which fixes an issue)

# Checklist:

<!-- delete when not applicable to your PR -->

- I have added tests
- I have made corresponding changes to the existing documentation
- I have created new documentation