L1-231: Rework build-aleph-e2e-client workflows (#1744)
# Description

Motivation of this PR is to fix `PR commit` workflow concurrency. As it
is now, when new commit is pushed to PR and there's a current pipeline
in progress, job `build-aleph-e2e-client` prevents concurrency to run
correctly, as concurrency only cancels the first level jobs in a
workflow. Currently, there's `check-vars-and-secrets` that is a
requirement for `build-aleph-e2e-client`. With this PR, job dependency
in `PR commit` workflow is cleaner, and `concurrency` works as expected.

Additionally, we opt out from pushing `aleph-e2e-client` to ECR as it is
not needed.

## Type of change

Please delete options that are not relevant.

- Bug fix (non-breaking change which fixes an issue)

# Checklist:

* [ Nightly pipeline logic e2e tests
](https://github.com/Cardinal-Cryptography/aleph-node/actions/runs/9204808809)
- PASS
* [ Nightly pipeline normal session e2e tests
](https://github.com/Cardinal-Cryptography/aleph-node/actions/runs/9204867831)
- PASS
* [ Nightly pipeline e2e tests on featurenet
](https://github.com/Cardinal-Cryptography/aleph-node/actions/runs/9204076309/job/25317637112)
- PASS