A0-4616: Don't request blocks you started iimporting (#1950)
# Description

By optimistically assuming that a block that we started importing will
get imported, we treat it as already imported for purposes of sending
requests. This should limit both the amount of received blocks, as well
as the amount of blocks in the import queue, both of which were slowing
sync down.

## Type of change

- Bug fix (non-breaking change which fixes an issue)
- New feature (non-breaking change which adds functionality)

# Checklist:

<!-- delete when not applicable to your PR -->

- I have added tests
- I have made corresponding changes to the existing documentation
- I have created new documentation